# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.BatchCreateBillsRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/handlers.CreateBillRequest'
        type: array
    type: object
  handlers.BillResponse:
    properties:
      amount:
        type: string
      category_id:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      description:
        type: string
      id:
        type: integer
      occurred_at:
        type: string
      type:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  handlers.CreateBillRequest:
    properties:
      amount:
        type: string
      category_id:
        type: integer
      category_name:
        type: string
      currency:
        type: string
      description:
        type: string
      occurred_at:
        type: string
      type:
        type: integer
    type: object
...
  handlers.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      success:
        type: boolean
    type: object
...
  handlers.UpdateBillRequest:
    properties:
      amount:
        type: string
      category_id:
        type: integer
      category_name:
        type: string
      currency:
        type: string
      description:
        type: string
      occurred_at:
        type: string
      type:
        type: integer
    type: object
...
info:
  contact: {}
  description: Accounting app backend API for auth, categories, transactions, stats,
    and API keys.
  title: Sumi API
  version: 0.1.0
paths:
...
  /api/bills:
    get:
      parameters:
      - default: 20
        description: Page size
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      - description: 'Transaction type: 1 expense, 2 income'
        in: query
        name: type
        type: integer
      - description: Category ID
        in: query
        name: category_id
        type: integer
      - description: Currency code
        in: query
        name: currency
        type: string
      - description: Case-insensitive substring match on description
        in: query
        name: keyword
        type: string
      - description: Start datetime (RFC3339)
        in: query
        name: start_time
        type: string
      - description: End datetime (RFC3339)
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.BillResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: List transactions
      tags:
      - Transactions
    post:
      consumes:
      - application/json
      parameters:
      - description: Transaction payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateBillRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.BillResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Create transaction
      tags:
      - Transactions
  /api/bills/{id}:
    delete:
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Delete transaction
      tags:
      - Transactions
    get:
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BillResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get transaction
      tags:
      - Transactions
    put:
      consumes:
      - application/json
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      - description: Transaction payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateBillRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BillResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Update transaction
      tags:
      - Transactions
  /api/bills/batch:
    post:
      consumes:
      - application/json
      parameters:
      - description: Transactions payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.BatchCreateBillsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/handlers.BillResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Create multiple transactions atomically
      tags:
      - Transactions
...
  /api/transactions:
    get:
      parameters:
      - default: 20
        description: Page size
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      - description: 'Transaction type: 1 expense, 2 income'
        in: query
        name: type
        type: integer
      - description: Category ID
        in: query
        name: category_id
        type: integer
      - description: Currency code
        in: query
        name: currency
        type: string
      - description: Case-insensitive substring match on description
        in: query
        name: keyword
        type: string
      - description: Start datetime (RFC3339)
        in: query
        name: start_time
        type: string
      - description: End datetime (RFC3339)
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.BillResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: List transactions
      tags:
      - Transactions
    post:
      consumes:
      - application/json
      parameters:
      - description: Transaction payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateBillRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.BillResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Create transaction
      tags:
      - Transactions
  /api/transactions/{id}:
    delete:
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Delete transaction
      tags:
      - Transactions
    get:
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BillResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get transaction
      tags:
      - Transactions
    put:
      consumes:
      - application/json
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      - description: Transaction payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateBillRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BillResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Update transaction
      tags:
      - Transactions
  /api/transactions/batch:
    post:
      consumes:
      - application/json
      parameters:
      - description: Transactions payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.BatchCreateBillsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/handlers.BillResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Create multiple transactions atomically
      tags:
      - Transactions
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: Programmatic API key for server-to-server access.
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'Use the format: Bearer {access_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"